Allow the tests to restart the Control Gallery if it crashes
