+++ Xamarin.Forms.Controls.Issues/Xamarin.Forms.Controls.Issues.Shared/TestPages/TestPages.cs
@@ -289,7 +289,7 @@ namespace Xamarin.Forms.Controls
 	}
 #endif
 
-	public abstract class TestPage : Page
+		public abstract class TestPage : Page
 	{
 #if UITEST
 		public IApp RunningApp => AppSetup.RunningApp;
