+++ Xamarin.Forms.Core.Windows.UITests/Xamarin.Forms.Core.Windows.UITests.csproj
@@ -74,6 +74,7 @@
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="WindowsTestBase.cs" />
+    <Compile Include="WinDriverApp.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
